refactor: move auth drawers and invalid-token-toast to proper place
