+++ src/app/forgot-password/page.tsx
@@ -6,7 +6,7 @@ import { Button } from '@/components/ui/button';
 import { getPasswordResetTokenDataByToken } from '@/utils/prisma-helpers';
 import { ResetPasswordForm } from '@/components/forgot-password/ResetPasswordForm';
 import { isTokenValid } from '@/utils/tokens';
-import { InvalidTokenToast } from '@/components/forgot-password/InvalidTokenToast';
+import { InvalidTokenToast } from '@/components/ui/InvalidTokenToast';
 import { SERVER_ERROR_MESSAGES } from '@/constants/server-error-messages';
 
 interface IForgotPasswordPageProps {
