+++ src/components/ui/header.tsx
@@ -18,7 +18,7 @@ import {
 } from '@/components/ui/dropdown-menu';
 import { NameEditionDrawer } from '@/components/ui/drawers/names/NameEditionDrawer';
 import { DeleteConfirmationDrawer } from '@/components/ui/drawers/deletions/with-checkbox/DeleteConfirmationDrawer';
-import { LogOutDrawer } from '@/components/ui/drawers/LogOutDrawer';
+import { LogOutDrawer } from '@/components/ui/drawers/auth/LogOutDrawer';
 import { useState } from 'react';
 import { DROPDOWN_ITEMS } from '@/constants/dropdown-items';
 import Link from 'next/link';
